Include audio level in Sample (RFC6464)
This allows Element's to get the audio level info, and hence to indicate
who is speaking to the signalling server, which can tell the clients.

This only really works if ion-avp is embedded in the signalling
server (used as a library), which isn't how ion-avp is intended
to be used. The correct way to do this is in ion-sfu.